backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
tweepy==4.14.0
pytrends==4.9.2
requests==2.31.0
python-multipart==0.0.6
pydantic==1.10.12
cors==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
nltk==3.9.0
newsapi-python==0.2.7
aiohttp==3.8.5
prophet==1.1.4
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
joblib==1.3.2
reportlab==4.0.7
